Move DiscoveryClient creation so it is guarded.
Move it back to EurekaDiscoveryClientConfiguration.

fixes gh-578
