+++ spring-cloud-netflix-core/src/main/java/org/springframework/cloud/netflix/eureka/EurekaClientAutoConfiguration.java
@@ -18,36 +18,20 @@ package org.springframework.cloud.netflix.eureka;
 
 import static org.springframework.cloud.util.IdUtils.getDefaultInstanceId;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import lombok.SneakyThrows;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
-import org.springframework.boot.autoconfigure.condition.AllNestedConditions;
-import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.SearchStrategy;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
 import org.springframework.cloud.client.CommonsClientAutoConfiguration;
 import org.springframework.cloud.client.actuator.HasFeatures;
-import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.client.discovery.noop.NoopDiscoveryClientAutoConfiguration;
-import org.springframework.cloud.context.scope.refresh.RefreshScope;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.ConfigurableEnvironment;
 
@@ -59,6 +43,8 @@ import com.netflix.discovery.EurekaClientConfig;
 
 /**
  * @author Dave Syer
+ * @author Spencer Gibb
+ * @author Jon Schneider
  */
 @Configuration
 @EnableConfigurationProperties
@@ -108,90 +94,4 @@ public class EurekaClientAutoConfiguration {
 		return new InstanceInfoFactory().create(config);
 	}
 
-	@Bean
-	public DiscoveryClient discoveryClient(EurekaInstanceConfig config,
-			EurekaClient client) {
-		return new EurekaDiscoveryClient(config, client);
-	}
-
-	@Configuration
-	@ConditionalOnMissingRefreshScope
-	protected static class EurekaClientConfiguration {
-
-		@Autowired
-		private ApplicationContext context;
-
-		@Bean(destroyMethod = "shutdown")
-		@ConditionalOnMissingBean(value = EurekaClient.class, search = SearchStrategy.CURRENT)
-		@SneakyThrows
-		public EurekaClient eurekaClient(ApplicationInfoManager applicationInfoManager,
-				EurekaClientConfig config, EurekaInstanceConfig instance) {
-			applicationInfoManager.initComponent(instance);
-			return new CloudEurekaClient(applicationInfoManager, config, this.context);
-		}
-	}
-
-	@Configuration
-	@ConditionalOnRefreshScope
-	protected static class RefreshableEurekaClientConfiguration {
-
-		@Autowired
-		private ApplicationContext context;
-
-		@Bean(destroyMethod = "shutdown")
-		@ConditionalOnMissingBean(value = EurekaClient.class, search = SearchStrategy.CURRENT)
-		@SneakyThrows
-		@org.springframework.cloud.context.config.annotation.RefreshScope
-		public EurekaClient eurekaClient(ApplicationInfoManager applicationInfoManager,
-				EurekaClientConfig config, EurekaInstanceConfig instance) {
-			applicationInfoManager.initComponent(instance);
-			return new CloudEurekaClient(applicationInfoManager, config, this.context);
-		}
-
-	}
-
-	@Target({ElementType.TYPE, ElementType.METHOD})
-	@Retention(RetentionPolicy.RUNTIME)
-	@Documented
-	@Conditional(OnMissingRefreshScopeCondition.class)
-	@interface ConditionalOnMissingRefreshScope {
-
-	}
-
-	@Target({ElementType.TYPE, ElementType.METHOD})
-	@Retention(RetentionPolicy.RUNTIME)
-	@Documented
-	@Conditional(OnRefreshScopeCondition.class)
-	@interface ConditionalOnRefreshScope {
-
-	}
-
-	private static class OnMissingRefreshScopeCondition extends AnyNestedCondition {
-
-		public OnMissingRefreshScopeCondition() {
-			super(ConfigurationPhase.REGISTER_BEAN);
-		}
-
-		@ConditionalOnMissingClass("org.springframework.cloud.context.scope.refresh.RefreshScope")
-		static class MissingClass {
-		}
-
-		@ConditionalOnMissingBean(RefreshScope.class)
-		static class MissingScope {
-		}
-
-	}
-
-	private static class OnRefreshScopeCondition extends AllNestedConditions {
-
-		public OnRefreshScopeCondition() {
-			super(ConfigurationPhase.REGISTER_BEAN);
-		}
-
-		@ConditionalOnClass(RefreshScope.class)
-		@ConditionalOnBean(RefreshScope.class)
-		static class FoundScope {
-		}
-
-	}
 }
